refactor(oracles/data_providers_sdk): Move stock market fetchers
