+++ libs/data_providers_sdk/src/price_data/fetchers/mod.rs
@@ -1,1 +1,2 @@
 pub mod exchanges;
+pub mod stock_markets;
