+++ libs/data_providers_sdk/src/price_data/fetchers/stock_markets/mod.rs
@@ -1,4 +1,5 @@
 pub mod alpha_vantage;
 pub mod fmp;
 pub mod twelvedata;
+pub mod utils;
 pub mod yahoo_finance;
