+++ libs/data_providers_sdk/src/price_data/fetchers/stock_markets/twelvedata.rs
@@ -5,12 +5,10 @@ use futures::{future::LocalBoxFuture, FutureExt};
 
 use serde::Deserialize;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
 
-use crate::utils::print_missing_network_price_data;
+use crate::price_data::fetchers::stock_markets::utils::print_missing_network_price_data;
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
 
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
 pub struct PriceData {
